CMDB/asset/requirements.txt
configparser==3.7.3
numpy==1.16.2
pandas==0.24.2
py-dmidecode==0.0.2
